django_ocr/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
google-cloud-documentai==2.24.0
google-cloud-vision==3.7.0
Pillow==10.2.0
PyPDF2==3.0.1
pdf2image==1.17.0
python-dotenv==1.0.1
mysqlclient==2.2.1
gunicorn==21.2.0
openai==1.57.4
